% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optfunLBSPRDome.R
\name{optfunLBSPRDome}
\alias{optfunLBSPRDome}
\title{Likelihood function for estimation model}
\usage{
optfunLBSPRDome(
  tryFleetPars,
  fixedFleetPars,
  LenDat,
  StockPars,
  SizeBins = NULL,
  mod = c("GTG", "LBSPR")
)
}
\arguments{
\item{tryFleetPars}{A list of fleet parameters varied as part of optimisation process.}

\item{fixedFleetPars}{A list of fixed fleet parameter, e.g. selectivity-at-length parameters.}

\item{LenDat}{Observed fish counts in each length bin in an integer vector.}

\item{StockPars}{Stock life-history parameters in a list.}

\item{SizeBins}{A list containing ToSize and Linc (bin width) for specifying length bins.}

\item{mod}{Specification of simulation mode: either GTG or original LBSPR.}
}
\value{
NLL
}
\description{
Calculates Negative Log Likelihood of length composition data
}
\details{
Calculates: 

* Predicted probability/proportions of fish caught in fishery in each length bin 
for the estimated fishing parameters from the simulation routine.
* The negative log likelihood (multinomial) of the observed data for those fishing parameters.

using -sum(LenDat * log(predProb/LenProb)). Penalises SL50/Linf values close to 1 using beta distribution.
}
\seealso{
* [optLBSPRDome()] optimises the likelihood (this function) of data by varying fishing parameters to obtain LBSPR fit
* [simLBSPRDome()] simulates per-recruit numbers-at-length in catch.
}
\author{
K. Hommik, C. J. Fitzgerald
}
