% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optLBSPRDome.R
\name{optLBSPRDome}
\alias{optLBSPRDome}
\title{Parameter estimation/optimisation model}
\usage{
optLBSPRDome(
  StockPars,
  fixedFleetPars,
  LenDat,
  SizeBins = NULL,
  mod = c("GTG", "LBSPR")
)
}
\arguments{
\item{StockPars}{Stock life-history parameters in a list.}

\item{fixedFleetPars}{The list of fixed fleet parameter, e.g. selectivity-at-length parameters.}

\item{LenDat}{Observed fish counts in each length bin in an integer vector.}

\item{SizeBins}{A list containing ToSize and Linc (bin width) for specifying length bins.}

\item{mod}{Specification of simulation mode: either GTG or original LBSPR.}
}
\value{
A list containing: 
\describe{
\item{\code{lbPars}}{Optimised fishing parameters.}
\item{\code{lbStdErrs}}{Standard errors of optimised parameters.}
\item{\code{fixedFleetPars}}{Fixed fishing parameters.}
\item{\code{PredLen}}{Predicted proportions-at-length in catch.}
\item{\code{NLL}}{Nonlinear log-likelihood from optimisation.}
\item{\code{MLE}}{Maximum likelihood estimates (log transformed values).}
}
}
\description{
Optimises fishing parameters by minimising NLL of length data
}
\details{
Calculates the initial estimates of fishing parameters for nonlinear optimisation:
 * \code{log(F/M) = log(0.5)}
 * \code{log(SL50/Linf) = log(Lmax/Linf)}
 * \code{log((SL95-SL50)/Linf = log(0.2*Lmax/Linf)}
and runs optimisation routine \code{optim} with these initial estimates. By setting \code{hessian = TRUE} 
we can calculate the hessian, its inverse the variance-covariance matrix of the estimates and
delta method approximations to standard errors.

Single parameter optimisation log(F/M) using the "Brent" method supports the specification of 
lower and upper bound constraints. Multiple parameter estimation through the "BFGS" method 
does not support bound constraints. 

Control parameters are specified through a control list and include
\describe{
\item{\code{maxit}}{Maximum number of iterations in optimisation.} 
\item{\code{reltol}}{Relative convergence tolerance.}
\item{\code{REPORT}}{Frequency of reports for the BFGS and other methods.}
\item{\code{trace}}{Non-negative integer 0 - no tracing, 1-6 higher levels produce more tracing information.}
}
}
\seealso{
* [optfunLBSPRDome()] calculates negative log likelihood of data given per-recruit model prediction.
}
\author{
K. Hommik, C. J. Fitzgerald
}
