% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLBSPRDome.R
\name{simLBSPRDome}
\alias{simLBSPRDome}
\title{LBSPR-dome simulation model}
\usage{
simLBSPRDome(StockPars, FleetPars, SizeBins = NULL)
}
\arguments{
\item{StockPars}{Stock life-history parameters in a list.}

\item{FleetPars}{A list of fishing fleet parameters: fishing mortality and selectivity-at-length.}

\item{SizeBins}{List specifying how to create size bins including Linc length increment.}
}
\value{
Output containing SPR, YPR, catch-length, population-length compositions for fished and unfished stocks
}
\description{
Per recruit simulation of exploited stock
}
\details{
Simulates length composition, SPR, YPR of exploited stock 
based on inputs including:
\describe{
\item{\code{StockPars}}{Life history parameters: growth, maturity, natural mortality.}
\item{\code{FM}}{Relative fishing mortality F/M.}
\item{\code{SL50, SL95}}{Fishing gear selectivity-at-length parameters.}
\item{\code{SizeBins}}{Specifications for a set of length bins.}
} 
Per recruit theory is used to derive length compositions based 
on the specified set of length bins.

Based on original GTG-LBSPR code developed by A. Hordyk 
accessible in <https://github.com/AdrianHordyk/GTG_LBSPR>. The 
original code was modified to include dome-shaped selectvity.
Details of modification in Hommik et al. (2020) accessible at
<https://doi.org/10.1016/j.fishres.2020.105574> User can pre-specify 
selectivity parameters based on dome-shaped parameterisations from 
SELECT method (Millar & Fryer, 1999). Minimum landing limit (MLL) 
can be defined. If MLL is defined then model assumes that selection      
probability for lengths < MLL is 0.
}
\seealso{
* [optLBSPRDome()] optimises the likelihood of data by varying fishing parameters to obtain LBSPR fit.
* [optfunLBSPRDome()] calculates negative log likelihood of data given per-recruit model prediction.
}
\author{
K. Hommik, C. J. Fitzgerald
}
